+++ src/Archs/MIPS/Zarem.Emulator.MIPS/Models/Enum/MIPSTrap.cs
@@ -1,6 +1,6 @@
 ﻿// Avishai Dernis 2026
 
-namespace Zarem.Emulator.Executor.Enum;
+namespace Zarem.Emulator.Models.Enum;
 
 /// <summary>
 /// An enum describing the kind of trap that occurred during an <see cref="Execution"/>.
