+++ src/Archs/MIPS/Zarem.Emulator.MIPS/Models/InstructionServiceTable.Co0.cs
@@ -1,13 +1,13 @@
 ﻿// Avishai Dernis 2026
 
 using System;
-using Zarem.Emulator.Executor.Enum;
+using Zarem.Emulator.Models.Enum;
 using Zarem.Emulator.Machine.Enums;
 using Zarem.Models.Instructions;
 using Zarem.Models.Instructions.Enums.Registers;
 using Zarem.Models.Instructions.Enums.SpecialFunctions.CoProc0;
 
-namespace Zarem.Emulator.Executor;
+namespace Zarem.Emulator.Models;
 
 public partial class InstructionServiceTable
 {
