+++ src/Archs/MIPS/Zarem.Emulator.MIPS/Models/InstructionServiceTable.Float.cs
@@ -2,13 +2,13 @@
 
 using System;
 using System.Numerics;
-using Zarem.Emulator.Executor.Enum;
+using Zarem.Emulator.Models.Enum;
 using Zarem.Models.Instructions;
 using Zarem.Models.Instructions.Enums;
 using Zarem.Models.Instructions.Enums.SpecialFunctions.FloatProc;
 using static Zarem.Emulator.Machine.CoProcessors.FloatProcessor;
 
-namespace Zarem.Emulator.Executor;
+namespace Zarem.Emulator.Models;
 
 public partial class InstructionServiceTable
 {
