+++ tests/Archs/MIPS/Test.Emulator.MIPS/ExecutionTests.cs
@@ -8,10 +8,10 @@ using Zarem.Assembler.Parsers;
 using Zarem.Assembler.Tokenization;
 using Zarem.Emulator;
 using Zarem.Emulator.Config;
-using Zarem.Emulator.Executor.Enum;
 using Zarem.Emulator.Machine;
 using Zarem.Emulator.Machine.Enums;
 using Zarem.Emulator.Machine.Registers;
+using Zarem.Emulator.Models.Enum;
 using Zarem.Emulator.TrapHandlers;
 using Zarem.Models.Instructions;
 using Zarem.Models.Instructions.Enums.Registers;
